[Tidy] Restructure vizro-ai (#561)
