+++ vizro-ai/src/vizro_ai/plot/_utils/_safeguard.py
