+++ vizro-ai/src/vizro_ai/plot/components/_base.py
