+++ vizro-ai/src/vizro_ai/plot/components/dataframe_craft.py
@@ -14,8 +14,8 @@ except ImportError:  # pragma: no cov
 from langchain_core.language_models.chat_models import BaseChatModel
 
 from vizro_ai.chains._chain_utils import _log_time
-from vizro_ai.components import VizroAiComponentBase
-from vizro_ai.schema_manager import SchemaManager
+from vizro_ai.plot.components import VizroAiComponentBase
+from vizro_ai.plot.schema_manager import SchemaManager
 
 logger = logging.getLogger(__name__)
 
@@ -126,7 +126,7 @@ class GetDataFrameCraft(VizroAiComponentBase):
 if __name__ == "__main__":
     import vizro.plotly.express as px
 
-    from vizro_ai.chains._llm_models import _get_llm_model
+    from vizro_ai._llm_models import _get_llm_model
 
     llm_to_use = _get_llm_model()
     df = px.data.gapminder()
