+++ vizro-ai/src/vizro_ai/plot/task_pipeline/_pipeline_manager.py
@@ -1,8 +1,8 @@
 """Pipeline Manager."""
 
 from langchain_core.language_models.chat_models import BaseChatModel
-from vizro_ai.components import GetChartSelection, GetCustomChart, GetDataFrameCraft, GetVisualCode
-from vizro_ai.task_pipeline._pipeline import Pipeline
+from vizro_ai.plot.components import GetChartSelection, GetCustomChart, GetDataFrameCraft, GetVisualCode
+from vizro_ai.plot.task_pipeline._pipeline import Pipeline
 
 
 class PipelineManager:
