+++ vizro-ai/tests/unit/vizro-ai/plot/components/test_explanation.py
@@ -1,6 +1,6 @@
 import pytest
 from langchain_community.llms.fake import FakeListLLM
-from vizro_ai.components import GetCodeExplanation
+from vizro_ai.plot.components import GetCodeExplanation
 
 
 @pytest.fixture
